Large file support for Win32: _stati64
the original par2cmdline-0.4 sources were not able to process
files larger than 2GB on the Win32 platform because diskfile.cpp used
the stat() function which only returns a signed 32-bit number on
Win32. This was changed to use _stati64() which returns a proper
64-bit file size. Note that the FAT32 file system from the Windows 95
era does not support files larger than 1 GB so this change is really
applicable only to files on NTFS disks - the default file system on
Windows 2000/XP/Vista.

(importer's note: par2-cmdline CVS fixed this on 2004-06-23, but never
had an official release including it.)