resync with CVS, update TBB, RS inner loop tweak, -d<directory> recursive support
synchronised the sources with the version of par2cmdline in the CVS at
sourceforge.net/projects/parchive

built against the 20070927 version of the Intel TBB

tweaked the inner loop of the Reed Solomon code so that the compiler
will produce faster/better/smaller code (which may or may not speed up
the program).

added support for creating and repairing data files in directory trees
via the new -d<directory> command line switch. See the
README_FIRST.txt file in the distribution for details on how to use
this.

(importer's note: README_FIRST.txt in the tarball for this version had
DOS line endings.  Fixed so the whole file doesn't show as rewritten.)