read files in order, throttle console, portability fixes
changed several concurrent loops from using TBB's parallel_for to
parallel_while so that files will be processed in a sequential (but
still concurrent/threaded) manner. For example, 100 files were
previously processed on dual core machines as:

Thread 1: file 1, file 2, file 3, ..., file 50

Thread 2: file 51, file 52, file 53, ..., file 100

which caused hard disk head thrashing. Now the threads will process
the files from file 1 to file 100 on a first-come-first-served basis.

--

limited the rate at which cout was called to at most 10 times per
second.

when building for i386 using GCC, this version will now build with an
assembler version of the inner Reed-Solomon loop because the code
generated by GCC was not as fast/small as the Visual C++ version.
Doing this should bring the GCC-built (POSIX) version's speed up to
that of the Visual C++ (Windows) version.

 for canonicalising paths on POSIX systems, the program will now try
to use the realpath() API, if it's available, instead of the fragile
code in the original version.

 on POSIX systems, attempting to use a parameter of "-d." for par2
creation would cause the program to fail because it was not resolving
a partial path to a canonical full path. Fixed.