critical bugfix for reedsolomon-inner-i386-posix.s; FreeBSD port
   the Linux version wasn't working because it was not built
correctly: the reedsolomon-inner-i386-posix.s was using an incorrect
include directive. Fixed.
    *** WARNING ***
        A consequence of this error is that par2 files created with
the 20080116 Linux binary contain incorrect repair data and therefore
cannot be used to repair data files. The par2 files will need to be
created again using either the 20071128 build of the Linux binary or
this build of it.
    *** WARNING ***

tweaked the Makefile and par2cmdline.h to allow for building
under FreeBSD.

first release of 32-bit and 64-bit binaries for FreeBSD (built under
RELEASE 6.2).

updated to use the 20080115 version of the Intel TBB library.