add -t option to choose what's done in parallel; dup removal of input filenames; realpath bugfix
added the -t0 option to allow verification to be done serially but
still perform repair concurrently, and for creation, MD5 checksumming
will be done serially and par2 data creation will be done
concurrently. The default is to perform all operations concurrently,
so if you want the new behaviour, you will need to manually specify
-t0 on the command line or build your own custom version of the
executable.

 if the realpath() API returned NULL, the par2 files created would end
up with the name of the first file in the list of files to create par2
files for. Fixed.

 no longer includes duplicate file names in the list of files to
create redundancy data for (which would otherwise bloat the .par2
files).

 now displays the instruction set being executed.

  updated to use the tbb20_017oss_src.tar.gz version of the Intel TBB
library.

 first release of a 64-bit Windows binary.