MMX support; build updates; async i/o; -0 option
 added more information to a few of the error messages to make it
easier to specify block counts, etc. when using the -d option.

 redundancy can now be specified using floating point values instead
of integral values, eg, 8.5% instead of 8% or 9%.

 added the -0 option to create dummy par2 files. This was done so that
the actual size of the par2 files can be quickly determined. For
example, suppose you wish to fill up a CD-R's or DVD-R's remaining
empty space with par2 files of the files filling up the disc, then by
using the -0 option, you can quickly work out whether the par2 files
will fit and by how much, which in turn allows you to maximize the use
of the remaining empty space (you would alter the block count number
and/or size so that the optimal number of blocks are created to fill
up the remaining space). To determine how much CD-R or DVD-R space you
have to fill, find out how many blocks your blank disc has (using a
burning program such as ImgBurn [Windows]) and how many blocks your
data would occupy when burned (using an image creation program such as
mkisofs [all platforms] which has a handy -print-size option). ImgBurn
[Windows] can also tell you how many blocks you have for filling if
you use its 'build' command. WARNING: be careful when using this
command that you don't burn the dummy par2 files that it creates
because they don't have any valid data in them. Remember, they are
created only to determine the actual size of the real par2 files that
would be created if you had not used the -0 option.

 added MMX-based code from Paul Houle's phpar2_12src version of
par2cmdline-0.4. As a result, the repair and creation of par2 files
using x86 or x86_64 MMX code is about 20% faster than the scalar
version (the scalar version is used if your computer is not MMX
capable). MMX CPUs: Intel Pentium II and later, AMD Athlon64 and later.

 added asynchronous I/O for platforms that support such I/O: Mac OS X,
Windows, GNU/Linux. This results in a small (~1-5%) improvement in
throughput, especially for repairing. Unfortunately, using async I/O
causes a crash under FreeBSD, so the pre-built binaries are built to
only use synchronous I/O.

 first release of 32-bit and 64-bit PowerPC binaries for Mac OS X. The
32-bit version requires at least 10.4, and the 64-bit version requires
at least 10.5. The 64-bit version is UNTESTED (because of lack of
access to a G5 Mac).

 first release of a 64-bit x86_64 binary for GNU/Linux. Tested under
the 64-bit version of Gentoo 2008.0.

 the 64-bit Windows binary is built using the tbb20_20080408oss
release of the TBB; the Mac, GNU/Linux, FreeBSD and 32-bit Windows
binaries are built using the tbb21_009oss release of the TBB. The
tbb21_009oss release does not support the VC7.1 runtime libraries on
Win64 so it was necessary to fallback to a previous version for the
Windows 64-bit binary.