async I/O fixes; build updates
 asynchronous reading of a large number of small files would sometimes
not complete which caused the program to hang. Fixed by reverting to
synchronous reading (most of the benefit of async I/O is from async
writing so this change does not affect overall performance).

 some operating systems have limits on the number of open files which
was easily exceeded when a large number of small files are being
processed for par2 creation or for repair. Fixed by closing the source
files as soon as they are no longer needed to be opened (which is
determined by counting how many data blocks the file provides for
creation/repair).

 the 64-bit Windows binary is built using the tbb20_20080408oss
release of the TBB; the Mac, GNU/Linux, FreeBSD and 32-bit Windows
binaries are built using the tbb21_009oss release of the TBB. The
tbb21_009oss release does not support the VC7.1 runtime libraries on
Win64 so it was necessary to fallback to a previous version for the
Windows 64-bit binary.

 the pre-built FreeBSD binaries are built to only use synchronous I/O
(to avoid a crash).