Fix MIPS, PPC arch compilation error: cannot bind packed field
Equal to following distribution specific patches:
	debian/000_unidentified.diff (part of)
	fedora/par2cmdline-packed.patch
	gentoo/par2cmdline-0.4-letype.patch
	slackware/letype.patch

"The GCC structure packing extension cannot be used with non-POD classes.
This restriction doesn't seem to be applied consistently, and as a
result par2cmdline only FTBFS on big-endian systems, where leu16, leu32
and leu64 are defined as class types.  However, all the classes used to
define the PAR file formats need to be POD structs.  In particular they
must not have constructors or copy-assignment operators and all their
non-static data members must be public.
Secondly, if a member of a packed structure is a structure that wasn't
defined as packed, the packed-ness acts as a qualifier on the member's
type, so you can't bind an ordinary reference (const MD5Hash &) to an
MD5Hash member of a packed structure.  This can be remedied by defining
the MD5Hash class as packed too."
from http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=356767