thread-safety; logging ratelimit; -t parallel; avoid heap fragmentation; use std::unique_ptr; build updates
 when creating parity files, the main packet was not always being
written to the parity files when they were processed concurrently
because the main packet was not being safely appended to the list of
packets to output because a non-thread-safe data container was being
used. This bug would manifest when a large number of source files were
being processed. Fixed by using a thread-safe data container.

 when creating parity files, the "Opening: <file>" messages will only
be displayed for the first n source files, where n defaults to 200.
This restriction was added so that creating parity files for a large
number of source files would not cause a lot of scrolling which in
turn would make the processing take a long time. Use the new -z<n>
command line switch to set a different limit. Use -z0 to specify no
limit.

 verification of extra files is now performed concurrently if
requested to do so (previously they were always verified serially)

 the -t parameter can now include a positive integer value to restrict
the logical number of CPUs with which to process data with.

 in the Windows version, the program's CPU scheduling priority can now
be specified using the -p parameter

 the heap became fragmented during the verification of data files
because the checksum data buffer was allocated and deallocated for
each file verified, which resulted in the program's memory footprint
(aka its "working set") steadily increasing during the verification
phase. This would result in the 32-bit Windows version failing to
verify large data sets because it could not allocate verification data
buffers. To solve this, the checksum data buffer is no longer
allocated and deallocated for each file verified. Instead, a pool of
checksum objects is created and that pool of objects is then used and
re-used for verifying data files. The size of the pool matches the
number of logical CPUs which the program is asked to use. This change
benefits all versions of the program because by reducing heap
fragmentation, larger data sets can be processed using less virtual
memory.

 numerous small code changes were made to remove unnecessary string
copying. Such redundant copying would further fragment the heap as
well as use up memory for temporary strings which did not need to be
allocated in the first place.

 updated to Intel TBB 4.3 Update 1

 removed use of MAX_PATH or other fixed-size path buffers to avoid
buffer overflow errors

 the program failed to build under newer C++ standard libraries
because they no longer provide std::auto_ptr<T>. Fixed by either using
std::unique_ptr<T> (if available) or by providing our own version of
std::auto_ptr<T>.

(importer's note: GNU C++ provides auto_ptr<T> (deprecated), but not
std::auto_ptr.)

 the 32-bit Mac OS X version now requires 10.5 or later

 stopped building the FreeBSD version because the FreeBSD ports system
can now build the par2 program and TBB library without requiring any
changes to the sources of either and because it isn't possible to
build a "portable" version of the program, in the sense that the TBB
library cannot be in the same directory as the par2 executable - it
must be installed into /usr/lib/, and that is a job best left to the
FreeBSD ports system.